Split off zbstore.go file into object.go and realization.go
Updates #345
